AppShell to prevent hydration mismatches via centralized client-side init
