+++ src/app/calendar/page.tsx
@@ -5,8 +5,6 @@ import { useTranslation } from "react-i18next";
 import PageHeader from '@/components/PageHeader';
 import WIPSection from "@/components/WIPSection";
 
-import '@/i18n';
-
 export default function CalendarPage() {
   const { t } = useTranslation(['calendar', 'common']);
