+++ src/app/groups/page.tsx
@@ -5,8 +5,6 @@ import React from 'react';
 import { useTranslation } from "react-i18next";
 import WIPSection from "@/components/WIPSection";
 
-import '@/i18n';
-
 export default function GroupsPage() {
   const { t } = useTranslation(['groups', 'common']);
