+++ src/app/preferences/GeneralSettings.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Button, { ButtonVariant } from '@/components/Button';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 type Props = {
